Move rc_buffer to its own module
